[StentExp] FUSION data fields in MultiBeamForceField
Due to previous changes in the implementation, two Data fields of the
component were non-exclusive while they should have been. In the
previous implementation, the _isPlasticHugues field was used with the
meaning of "use the plasticity model with hardening, described in a
publication by Hugues" (Theoretical foundation for large scale
computations for nonlinear material behaviour, 1984). This implicit use
of hardening was contradictory with the existence of the
_isPerfectlyPlastic Data field, as both parameters could be 'true' at
the same time, not making any sense.

With this commit, we suppress the _isPlasticHugues field, to keep only
_isPerfectlyPlastic with the following interpretation:
  - If _isPerfectlyPlastic is true, then the plasticity model we used is
the one presented in Hugues' publication, for perfect plasticity. In our
implementation, it corresponds to three functions: computeElasticForce,
computePlasticForce, and computePostPlasticForce.
  - If _isPerfectlyPlastic is false, we apply the linear combination of
isotropic and kinematic hardening, presented in the same publication. In
our implementation, this correspond to the computeForceWithHardening
method.

In the process, we suppress an entire part of the computeStressIncrement
method, which was initially dedicated to a hardening model of
plasticity. In the previous state of the implementation, this part was
never accessed, as _isPerfectlyPlastic had to be true in order for
computeStressIncrement to be called (so on a perfect plasticty model).
The corresponding code is mostly prsent in the computeForceWithHardening
method, dedicated to hardening. This suppression sould not change
anything in the algorithm behaviour.