[StentExp] IMPLEMENT consistent tangent operator
Following the previous commit, we implement the computation of the
consistent tangent operator.
At the moment, the expression we use corresponds to the perfect
plasticity model, but it is implemented for both our perfect plasticity
model, and our mixed hardening model. This should be changed in future
commits, by adding the expression of the consistent tangent operator
when hardening is considered.
In addition, all the computation is made for the moment with the matrix
notation, in 9 dimensions. In future commits, we should consider
switching to Voigt notation, in order to reduce the computations.