[StentExp] CORRECT computation in perfect plasticity
Due to an old, obsolete implementation of the perfect plasticity model,
the computation in "post-plasticity" (i.e. when Gauss points went from a
plastic mode into an elastic mode) was not incremental. This was in
contradiction with the litterature we base our implementation on, and
incoherent with the nonlinearity of plastic models.

Accordingly, we changed our implementation of the perfectly plastic
case, in order for the computation to be entirely incremental. We
created two new functions, inspired from the one we use for our
hardening model.

At the moment, we keep in the implementation the old methods
computeElasticForce, computePlasticForce, and computePostPlasticForce,
but they are no longer used in the computations. In future commits,
these should be deleted.