[StentExp] CORRECT plastic hardening implementation
We cleaned the previous implementation of hardening, basing the new
calculation on the detailed work of Simo and Hugues in 'Computation
Inelasticity', 1998.
The bottomline is that the mixed linear model of hardening we use with
the Von Mises plastic criterion does not require an implicit routine to
solve for the new stress at each iteration.

For the moment, we let the previous implicit implementation in comments,
as it could be used for more complex types of hardening. It should,
however, be moved in a dedicated method in the future.