[SofaKernel] Add DataLink object & PathResolver.
Currently in Sofa the Link implementation is very hard to maintain because
in a single implementation based on type-traits we are trying to handle to object
of different nature that are BaseData and Base.

Link between Base is used everywhere in Sofa but a Link between two data
is only to hold the parent of a Data. It is questionable if such complex
implementation for Link is needed for just this single use case ?

Not doing that have a lot of consequence:
- Data implementation is simpler (because it is no needed to have Class )
- Link implementation is much simpler (and Link is used everywhere)
- The instanciation of Link is time consuming in the compilation process and
  requieres a lot of #include<> making the dependency graph more dense and thus
  slow to compile.

So in this PR I introduce a new DataLink & AbstractDataLink that aims at replacing
SingleLink<BaseData*, BaseData*>.

In this PR I also add PathResolver, as path link resolution is important an currently it is spread everywhere in the sofa codebase. The underlying idea
is to have all path related things in a single location so we can check there is no duplicated code.

The implementation are provided with unittest. The PR does not touch other files.
A second PR will be in charge of refactoring Data.h. But to minimize deltas I prefer to have
this one merged first.