[StentExp] REMOVE deprecated methods from BeamPlasticFEMForceField
A previous implementation of the stress computation was based on three
methods (computeElasticForce, computePlasticForce and
computePostPlasticForce), in order to differentiate a non-incremental
solution for purely elastic deformation, and an incremental solution for
plastic or partially plastic deformation.
This implementation was responsible for strong instabilities in the
simulation, as well as for a recovery of the normally permanent plastic
deformation. We replaced it with a fully incremental solution, based on
the radial return algorithm for Von Mises plasticity. With this commit,
we suppress the deprecated code, which was commented and should have no
repercussion on the simulation.