[StentExp] REORGANISE force computation code
We reorder the force computation code and delete a commented part in the
hardening force computation method. This commented part was design for
an implicit iterative computation of the stress increment, which is
useless for the radial return algorithm with Von Mises-Hill plasticity.
It could be reused if a different model of plasticity was to be
implemented.
Although the reorder of the methods in the .inl file is detected as an
important modification by Git, no uncommented part of the code was
actually changed.