[BeamPlastic] CHANGE plugin name
Files implementing the BeamPlasticFEMForceField component were initially
extracted from a different plugin dedicated to stent expansion. To keep
the corresponding Git history, we kept the original 'StentExp' plugin
name in previous commits.
Here, we change the plugin name to BeamPlastic and modify the source
files and CMakeLists files accordingly. No functional change is made
regarding the code of the BeamPlasticFEMForceField component with this
commit.