[SofaHelper] Separate Definition/Declaration/Specialization in vector.h
Should'nt break too much. As vector.h is still including everything.
I know that vector[integral] is unconventionnal...but I actually find it
very clear way to express the fact the a file is containing specialization
of a templated object.

Eg:
vector<T>  is in vector.h
while vector<string> specialisation are in vector[string].h

I was thinkin to use other convention to makes the specialization more saillants