[SofaCore] Remove "depend" from Node (#1763)
Remove sortComponent() and the _depend_ data field in Node.h
I'm not sure this feature is used at all but the way of thinking among sofa developper is that if you need to have your component defined in a specific correct ordering then this should be the consequence of having a correct dependency graph between link & data. If this removal breaks some of your code please contact sofa-framework so we can consider reverting that change.