[SofaGuiQt] ADD qt.conf.h for custom qt.conf loading at runtime
This is needed by the SofaPython3 modules using Qt.
When importing them directly with python3 interpreter, <sofa-path>/bin/qt.conf is not found because Qt looks next to the executable (python3).
With this new qt.conf.h and its function sofa::gui::qt::loadQtConfWithCustomPrefix(), it is possible to force Qt to load the qt.conf we want.

qt.conf.h is based on the output of `<qt-path>/bin/rcc qt.conf.qrc -o qt.conf.h` with
    <qt-path>/bin/rcc: Qt RCC executable
    qt.conf.qrc: file to create next to <sofa-path>/bin/qt.conf to overload Qt default search path ":/qt/etc/qt.conf"
        ```xml
        <!DOCTYPE RCC><RCC version="1.0">
          <qresource prefix="/qt/etc/">
            <file>qt.conf</file>
          </qresource>
        </RCC>
        ```
    <sofa-path>/bin/qt.conf: this file is not in the command but is needed because referenced by qt.conf.qrc

Usage example:
    ```cpp
    std::string sofaPrefixAbsolute = sofa::helper::Utils::getSofaPathPrefix();
    std::string inputFilepath = FileSystem::cleanPath(sofaPrefixAbsolute + "/bin/qt.conf");
    bool success = loadQtConfWithCustomPrefix(inputFilepath, sofaPrefixAbsolute);
    if(success)
    {
        msg_info("Sofa.Gui") << "Loaded qt.conf from " << inputFilepath << " customized with Prefix = " << sofaPrefixAbsolute;
    }
    else
    {
        msg_warning("Sofa.Gui") << "Failed loading and/or customizing qt.conf from " << inputFilepath;
    }
    ```