[All] CLEAN definitions of Module and Plugin
Plugin = Something that MUST be loaded by the PluginManager, through a RequiredPlugin for example. The entry point of a Plugin is initExternalModule (and that name could be better).

Module = Something that CANNOT be loaded by the PluginManager. Most of the time, modules are not loaded manually, they are loaded implicitly during dependency resolution. The entry point of a Module is init (and that name could also be better).

From a CMake point of view, both Plugin and Module can be standalone packages or simple targets added to a bigger package.