[BeamPlastic] REMOVE PoissonContainer and StiffnessContainer
Both containers should be used to implement non-constant Poisson ratio
and Young modulus over the considered domain. As it is not (yet) the
case for the BeamPlastic, we remove them for better code readability.

If these features would be implemented in the future, it would be better
to use the classes defined in :
modules\SofaGeneralSimpleFem\src\SofaGeneralSimpleFem.
Until now these classes were duplicated in the BeamPlastic plugin, which
anyway was an issue.