[SofaConstraint] Parallel free motion and collision detection (#2040)
* [SofaSimulationCore] Split collision in 3 steps

* [SofaConstraint] free motion and collision are concurrent

* [SofaConstraint] Clean parallelization

* [SofaConstraint] Avoid race conditions

* [SofaConstraint] Parallel computation is off by default

* [SofaConstraint] Cleaning

* [SofaConstraint] Cleaning

* Add a benchmark scene to compare effects of multithreading in FreeMotionAnimationLoop

On my computer, with the single threaded scene, FreeMotion+CollisionDetection runs in 24.92 ms in average. FreeMotion runs in 7.7 ms and CollisionDetection in 16.36 ms. If these two steps runs in parallel, we expect a gain of 7.7 ms (the minimum between 7.7 and 16.36).

On my computer, with the multi-threaded scene, FreeMotion+CollisionDetection runs in 16.93 ms in average. It is a gain of 7.99 ms (=24.92-16.93), which is about what we expected.

It validates the implementation of the parallelization of the collision detection and the free motion steps in the FreeMotionAnimationLoop.

* [SofaConstraint] Fix compilation after merge