[All] ADD option to fetch pull-request commits in ExternalProject (#1961)
When your SOFA PR fails on CI somewhere in an external project, it may be because your change is breaking for that external project.
In that case, the best (and preferred) solution is that you push a fix in the external project so that when your PR is merged in SOFA, it does not fail on CI.
Here is the typical workflow for that use case.

    You open a PR in SOFA (let's call it "SOFA PR").
    SOFA CI fails to build/test your SOFA PR because some external projects are enabled and you didn't enable them in your local tests.
    You realize that the failure is due to your changes.
    You fix the issue in the external project and open a PR (let's call it "external PR").
    You reference your external PR in the description of your SOFA PR with
    [ci-depends-on https://github.com/<org>/<repo>/pull/<id>]
    SOFA CI can now build/test your PR. It will automatically replace the external project with the "result of the merge of your external PR".
    As soon as SOFA CI is green and external PR is merged, your SOFA PR is ready to merge.

This way, you do not need to touch any ExternalProject.cmake file in your SOFA PR +1

The only change in this PR consists in adding GIT_CONFIG "remote.origin.fetch=+refs/pull/*:refs/remotes/origin/pr/*" to all ExternalProject.cmake files.
This change permits to have access to any commit in any pull-request (including merge commits). See this nice cheatsheet about how to fetch PRs.
All other changes have been done in CI scripts and are ready to use.