[GmshLoader] Update MSH format version detection
Instead of assuming that the MSH file format version is either 1 or 2,
we read it from the file in order to handle the most recent format (4.1
at the time). The version is then passed to the parser in MeshGmsh.cpp.
With this commit, versions >= 4 are not yet handled by the parser.