[GmshLoader] Add test
At the moment the test fails, because there seem to be some discrepancy
between the topology that the MeshGmshLoader is able to retrieve in a
scene, and the data it actually contains. This might come from the fact
that the doLoad() method of MeshGmshLoader doesn't call anymore
its method readGmsh, and that the mesh is in fact retrieved in the scene
at the MeshGmsh level. This has yet to be clarified.