[BeamPlastic] Add structs to handle differently the gaussian quadrature
The idea is mainly to store the plasticity variables associated to each
Gauss point (for instance the shape function matrix, the local yield
stress, ...) in a dedicated object (GaussPoint3) rather than in vectors
stored in BeamInfo. This way, we can remove the depency of the
plasticity model on BeamInfo, with the idea to then inherit all this
information from BeamFEMForceField, and leave in
BeamPlasticFEMForceField only code related to plasticity.