[BeamPlastic] Fix computation error in conversion
When converting a 4th order tensor such as the generalised Hooke's law
from a Voigt matrix notation (6x6) to a 'vect' matrix notation (9x9), we
have to divide by 2 the components associated to the symmetric terms of
the corresponding 2nd order tensors (for instance strain if we are
refering to the generalised Hooke's law). The reason is that these terms
are considered twice in a regular tensor product, but only once if the
Voigt notation is used.