Fix detection of .msh files with version 1.x
The first line of old .msh files (version = 1.x) seems to be "$NOD\r"
instead of "$NOD", which caused them not to be detected by the
MeshGmshLoader parser. This issue comes from the fact that before the
update of the loader, if the header line was not recognised, then the
version of the parsed .msh file was assumed to be 2.x.

This fix simply adds $NOD\r$ as a recognised header line for .msh files
with version 1.x, mainly to maintain compatibility with the unit tests.