Fix failing tests
Some tests in the CI use .msh files with slightly different syntax than
the standard. Namely, the problem comes from the presence of a carriage
return character "\r", sometimes preceeded by a space " " character, at
the end of main section lines such as "$MeshFormat \r", or "$Nodes\r".
A solution to this is to make the parser a little bit more permessive,
by only parsing the beginning of these lines and see if it corresponds
to the expected section titles ("$MeshFormat", "$Nodes"). This is only
implemented for versions <= 2, as it is not actually part of the
standard (i.e. a .msh file starting with "$MeshFormat \r" instead of
"$MeshFormat" is not technically a .msh file).
The issue could come from .msh files commited from different OS, with
lines ending differently. If so, the more permessive parsing should also
be extended for versions >= 4.