[Sofa.Type] Speed up fixed_array constructors with perfect forwarding (#2450)
I think constructors in fixed_array, Vec and Mat passed their arguments by value. Perfect forwarding avoids this copy. Also, constructors of base classes are called instead of re-defining a similar constructor.

Now the construction of a 3x3 matrix is 3x faster and close to the construction of a 3x3 matrix with Eigen.