[SofaBaseLinearSolver] Remove the IsRequired flag in CGLinearSolver
PR 2419 restores the correct behavior of CGLinearSolver regarding the call-super-init
https://github.com/sofa-framework/sofa/pull/2419

But as some of its data are tagged as Required, this rise en error message when the value
is not set by the user.

I'm not sure it is wise to use the required flag when data provides meaningful and usable default values.
So I recommand removing it.