[SofaKernel] Add a namespace like behavior to factory & "à la" python import
This PR aims at handling in a proper way the components implemented in different plugings.
The PR is composed of two aspects.

1) Registering components by "plugin.classname" in the Factory
The factory currently access component by their classname only causing
collision when two component have the same name/template.
To avoid this, the PR add the plugin name as a prefix for the Factory'key.
So as an exemple, a component created with <OglModel> in the past is now referenced
as <SofaOpenglVisual.OglModel>.

Backward compatbility is possible by hard coding into the factory registration the adding of an alias from
new names to old name;
Eg adding an alias saying:
   "OglModel" => "SofaOpenglVisual.OglModel"

This makes the change 100% compatible with existing sofa scenes.

2) Importing plugin and component in a clear way (à la python)
The PR add a new component (FromComponent) that import one or several components.

Examples of use:
	 # Only <OglModel> is accessible in the scene
	<From plugin="SofaOpenglVisual" import="OglModel"/>
	<OglModel/>

	# I like to have my own "custom" name in my scene because I ate "official ones"
	<From plugin="SofaOpenglVisual" import="OglModel" as="VisualModel"/>
	<From plugin="SofaOpenglVisual" import="OglModel" as="RenderingModel"/>
	<VisualModel/>
	<RenderingModel/>


	# I'm a super lazy man so I import everything
	<From plugin="SofaOpenglVisual" import="*"/>

Signed-off-by: Damien Marchal <damien.marchal@univ-lille1.fr>