[SofaSparseSolver] Introduce FillReducingOrdering (#2501)
A new DataEngine is introduced: `FillReducingOrdering`

A DataEngine to reorder the degrees of freedom in a mesh in order to reduce fill-in in sparse matrix factorization.

In other terms, the algorithm minimizes the number of non-zeros entries in the factorization of the sparse matrix of a FEM problem by reordering the degrees of freedom.

The implementation is based on METIS.

Note: some of the direct linear solvers embed such a reordering internally (e.g. SparseLDLSolver), but not all of them (e.g SparseLUSolver, SparseCholeskySolver etc).


An example is also introduced (`examples/Components/engine/FillReducingOrdering.scn`).
Here is the comparison of the two system matrices:

Initial matrix:
![initial_matrix00001](https://user-images.githubusercontent.com/10572752/143068764-2cbd09b4-906b-4bc1-b3c6-03cd3034a1ab.jpg)

Reordered matrix:
![reorder_matrix00001](https://user-images.githubusercontent.com/10572752/143068809-80a5172a-ce39-4c53-bbfb-e8ec12066cf8.jpg)

It improves a lot the performances when the linear solver does not embed a fill-in reducing reordering. This is the case of the `SparseLUSolver` in the example scene.
Rough framerate without reordering: **2.5 FPS**
Rough framerate with reordering: **25 FPS**