[all] Rename the forwarding helper in sofa::simulation to use the "official way" done in sofa::core
To reduce the #include drama it was decided to use forward declaratin as much
as possible combined with explicit up-casting and down-casting.

There was such mecanism in simulation/fwd.h.
When we decided to use the same mecanism in core/fwd.h some unified convention where selected.

This PR updates simulation/fwd.h so to match the selected convention and fix the rest of the code.