[ArticulatedSystemPlugin] Prevent duplicated call to init().
The init function can only be called on object that are not in the
ready state. Otherwise call reinit() or implement a data update mecanism.

Two components in ArticulatedXX are changed so we don't enter into init method if the object
is already in a ready state because of their implementation they are not working
properly in such case.