[Sofa.DefaultType] writeOpenglMatrix is defining GLdouble.
This is wrong in any way as we shouldn't have a dependency to opengl.
Either we need to really know GLdouble from open...in that case writeOpenGLMatrix
should become a free function moved to some ogl packager... or we actually don't care so much
and in that case we juste use "double" instead.