[Core] Dynamic control of the computation of the bounding box (#3080)
A user can chose in the scene to update the bounding box or not. It is no longer a compile-time selection.

In InextensiblePendulum.scn, the second pendulum explodes. An update of the bbox would adapt to the exploded object. The bbox is so large that the first pendulum is no longer visible. It may lead the user to think that the scene is not working properly.