[Sofa.DefaultType] Add a map to handle double registration scenario (windows dll duplicated static definition)
Because windows don't merge identical symbols in different dll, it is possible to have
the same type register to different "id". To avoid this situation and insure having a single
"id" for every type a map is added to check unicity by type_index(type_info) which is
a stable way to get a unique index for a data type.