[Sofa.Core] Implementing a new mecanism for template inference for component.
The existing mecanism to infer the templates when they are not specified in the scene
is based on the canCreate method failure/success. The selected template is the one the
first one that does not fail.

The existing system has several drawback:
- duplication of the different similar behavior for a lot of components...while implementation of a lof of custom policies.
- mixing the "constraints" an object must meet to be created and what should be the behavior when the template is not specified.

In the present PR I fix that by offer a clear API for defining per-component their
template deduction mecanism. The mecanism is not based on the in-herithance relationship which allows
to factorize the implementation and thus enforce a behavioral unification. In addition as the
implementation is no more in the .h this contribute to reduce the need of includes.