[Sofa.Core] Remove the use of Link in MechanicalParams & ConstraintParams (#3304)
* [Sofa.Core] Remove the use of Link in MechanicalParams & ConstraintParams

Each readXX method exists in two implementation. One that takes State* while the
other is taking a SingleLink<State>.

Given that from a SingleLink<State> one can get the corresponding State* using the get() method
it is questionnable why to maintain two set of methods for the same service.

To mimimize the surface of the API I removed the link based version and update the code base accordingly.

* [Sofa.Core] Add forward declaration for SingleLink.h

* [Sofa.Core] SOFA_ATTRIBUTE_DEPRECATED__READX_PARAMS

To handle the three step deprecation process.