Introduce CMake presets (#3305)
See doc https://cmake.org/cmake/help/latest/manual/cmake-presets.7.html

It gives 3 different presets similarly to what we are used to on the CI. The idea is the same (3 different config) but it cannot mimic it perfectly. For example, the CI checks if dependencies are available before setting a variable. It is not possible with presets.