[Config] Change type of CMake message when adding a module (#3381)
Previously, the default mode was NOTICE, which print the message in red. The doc says "Important message printed to stderr to attract user's attention.". The consequence was that almost all the CMake log was written in red, and it was difficult to find the true important messages. This commit change the type of message so it is written in black.