[VectorTypeInfo] Change default handling of data buffer
Default behaviour of the getValuePtr methods was to give an address to
the vector data by returning a reference to the vector first element.
However the case where the method is called on a valid empty vector is
not handled.
To change this, we use the std::vector.data() method, which handles the
case when the vector is actually empty (returning a pointer that may or
may not be null)