[Spring] Limit spring force to some axis (#3850)
* [ForceField] Add direction option in RestShapeSpringsFF

The idea of this commit is to add the ability for a user to select in
which directions the RestShapeSpringsFF will be effective during
simulation. For this we add a new data field containing 6 booleans: one
for each direction of effort (3 directions of translation, 3 of
rotation). For each entry, "true" indicates that the direction will be
taken into account, and "false" that it won't. The default value of the
parameter is set to [1 1 1 1 1 1], so that all directions are taken into
account if the parameter is not specifically provided by the user. This
way, we expect the component not to be breaking.

/!\ At the moment, we only implemented the restriction for translations,
but not for rotations. Concretely, changing one of the last 3 entries of
the data field to "false" doesn't have any effect on the simulation. The
implementation for rotations should be the object of further commits.

* [ForceField] Add test scene to test the new RestShapeSpringsFF option

* [ForceField] Extend direction option in RestShapeSpringsFF

Following previous commit, we extend the direction option to rotations.

* [ForceField] Update RestShapeSpringsFF test scene

* [ForceField] Fix warning in test scene

* [ForceField] Add callback in RestShapeSpringForceField

The callback is meant to recompute the internal variables of the
components if the input MechanicalObject indices are changed
dynamically.

* Make activeDirections more generic

* No need to be a Vec

* make description matches dimension

* Fix wrong size

* Set component state

* Fix scene

* Remove check of component state

This is going to be checked earlier (but in another pull request)

* Refactor setting the component state

It now follow the same pattern used in other components: the component state is set to valid only at the end

---------

Co-authored-by: Camille Krewcun <camille.krewcun@inria.fr>