[Sofa.Simulation] Remove Node::bwdInit
The bwdInit is used to implement a bugfix that seems to be fixed in a more approriate way
elegant way in BaseMapping
```cpp
bool BaseMapping::insertInNode( objectmodel::BaseNode* node )
{
    if( isMechanical() ) node->addMechanicalMapping(this);
    else node->addMapping(this);
    Inherit1::insertInNode(node);
    return true;
}
```

That said, I disagree on the two fixes taht are considering that a node is in either a the mapping list or the mechanicalmapping list.
This is not consistent with was Inheritance graph is saying... a mechanicalmapping is a mapping and thus should be in the two lists.