[Helper] Update ComponentChange for MechanicalMatrixMapper
and MappingGeometricStiffnessForceField

The message goes from:
```
Object type MechanicalMatrixMapper<Vec3d,Vec3d> was not created
The object 'MechanicalMatrixMapper' is not in the factory.
This component has been MOVED from SofaGeneralAnimationLoop to Sofa.Component.Mapping.MappedMatrix since SOFA v22.06.
To continue using this component you may need to update your scene by adding
<RequiredPlugin name='Sofa.Component.Mapping.MappedMatrix'/>
```

to:

```
The object 'MechanicalMatrixMapper' is not in the factory.
This component has been REMOVED since SOFA v23.06 (deprecated since v23.12).
Please consider updating your scene. If this component is crucial to you please report in a GitHub issue in order to reconsider this component for future re-integration.
```