[Sofa/MechanicalLoad/TrianglePressureForceField] Add buildStiffnessMatrix & deprecate non constant mode
The non constant computation of the pressure is not valid as all the derivatives are
zero in addDForce, addKtoMatrix... it is also not consistant with QuadPressureForceField.

Moving to the situation with a valid non constant computation on both Triangle and Quad requires more
work, so for the moment I prefer to just deprecate the non constant situation.