[Base] Fix a minor bug related to the printing of info message while parsing
Parsing emits some messages with msg_info(). But this one is controlled by the
printLog data field. The consequence is that every fields parsed before the "printLog" one
are not reporting messages properly. To fix it we need to process the printLog one before
any other so there is a consistent result.