[Base] Fix a minor bug related to the printing of info message while parsing & minor upgrade to c++11 (#5191)
* [Base] Fix a minor bug related to the printing of info message while parsing

Parsing emits some messages with msg_info(). But this one is controlled by the
printLog data field. The consequence is that every fields parsed before the "printLog" one
are not reporting messages properly. To fix it we need to process the printLog one before
any other so there is a consistent result.

* [Base] Minor code upgrade to iterate over maps using structured binding.

* [Base] Extra c++11 replacement of "for earch" loops.

* Update Sofa/framework/Core/src/sofa/core/objectmodel/Base.cpp

Co-authored-by: Paul Baksic <30337881+bakpaul@users.noreply.github.com>

---------

Co-authored-by: Paul Baksic <30337881+bakpaul@users.noreply.github.com>