[Sofa.Simulation] Change event propagation pruning (#5749)
Currently the pruning event is done at node level.
Despite this is not fundamentally wrong, this behavior sounds a bit weird to me as It
imply there is no event pruning between controllers's which kind of introduce kind of
implicit coupling of event management at node level.

The PR remove this coupling

Co-authored-by: Hugo <hugo.talbot@sofa-framework.org>